doc: update member list with ncu-team sync
Update with command `ncu-team sync README.md`

There is already a `Standards` team in the `openjs-foundation` github
org.

I believe this is from a prior collab summit but not 100%. I've added
members who had opened a PR, and there are a handful of other folks in
this update that were already members. I think if we want to remove any
folks we should have that as a separate conversation

This is to serve as an alternative to #44, #45, #46, #47, #48

Closes: #41